% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_weighted_estimate.R
\name{calculate_weighted_estimate}
\alias{calculate_weighted_estimate}
\title{Calculate Weighted Estimate}
\usage{
calculate_weighted_estimate(df, est, lb, ub)
}
\arguments{
\item{df}{The data frame as generated by \code{exhaust_design(..., weight = TRUE)}.}

\item{est}{A character value indicating the variable in \code{df} that contains the point estimate.}

\item{lb}{A character value indicating the variable in \code{df} that contains
the lower bound of the estimate.}

\item{ub}{A character value indicating the variable in \code{df} that contains
the upper bound of the estimate.}
}
\value{
A list containing the
\describe{
 \item{"Estimate"}{Weighted estimate}
 \item{"Lower bound"}{Weighted lower bound}
 \item{"Upper bound"}{Weighted upper bound}
 \item{"n"}{Number of model estimates that have been used to generate the weighted estimate}
}
}
\description{
Provide an estimate based on user-specified weights for the choices of the research design.
}
\details{
See the vignette of the package for further details on how to implement the RDF workflow.
}
\examples{
\dontrun{
  print("Sorry. No examples yet.")
}
}
