% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_design.R
\name{define_design}
\alias{define_design}
\title{Define research design and create code directory with step templates}
\usage{
define_design(steps, rel_dir = ".", one_file = FALSE,
  one_file_name = "design_steps.R")
}
\arguments{
\item{steps}{A character vector containing the names for the strp functions.
All members need to be valid R names.}

\item{rel_dir}{A relative path to a directory in which you want to create
the step templates.}

\item{one_file}{If \code{TRUE}, then all templates will be stored in a file
with the name \code{one_file_name}. If {FALSE} (the default), each step template
will be stored in a separate file with the step name acting as file name.}

\item{one_file_name}{The name of the code file conating the step templates if
all step templates are stored in one file. Defaults to "design_steps.R".}
}
\value{
The \code{steps} parameter.
}
\description{
This generates a \code{code} directory in \code{rel_dir}
(defaulting to current working directory). In this directory it will create
a set of files with templates for all code steps or a joint file
containing all step templates.
}
\details{
See the vignette of the package for further details.
}
\examples{
\dontrun{
  print("Sorry. No examples yet.")
}
}
