% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{doc}
\alias{doc}
\title{Output a series of characters with newlines}
\usage{
doc(...)
}
\arguments{
\item{...}{parameters passed on to \code{paste(..., sep ="\n")}}
}
\value{
The return value from \code{paste()}
}
\description{
This is a small helper function for documetation
}
\examples{
  doc("## Documentation",
    "Documentation withtout the need to",
    "include new lines is wonderful."
  )
}
