% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rdf_estimates.R
\name{plot_rdf_estimates}
\alias{plot_rdf_estimates}
\title{Plot Researcher Degrees of Freedom Estimates}
\usage{
plot_rdf_estimates(df, est, lb, ub, color = "blue", n = 30)
}
\arguments{
\item{df}{The data frame as generated by \code{exhaust_design()}.}

\item{est}{A character value to indicate the variable in \code{df} that contains the pont estimate
that you want to plot.}

\item{lb}{A character value to indicate the variable in \code{df} that contains
the lower bound of your estimate.}

\item{ub}{A character value to indicate the variable in \code{df} that contains
the upper bound of your estimate.}

\item{color}{Which color do you want your estimate density to have?}

\item{n}{The condidence interval is currently naively estimate by uniformly distributing \code{n}
observations between \code{lb} and \code{ub}}
}
\value{
A \code{ggplot} object containing the plot.
}
\description{
Visualizes the estimates generated by exhausting all documented researchers degrees of freedom.
}
\details{
Currently naively assumes a uniform distribution between lower and upper bound
  of the estimate to display the confidence interval.
  See the vignette of the package for further details on how to implement the RDF workflow.
}
\examples{
\dontrun{
  print("Sorry. No examples yet.")
}
}
