% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rdf_ridges_by_dchoice.R
\name{plot_rdf_ridges_by_dchoice}
\alias{plot_rdf_ridges_by_dchoice}
\title{Plot Estimate Ridge Lines by Discrete Choice}
\usage{
plot_rdf_ridges_by_dchoice(df, est, dchoice, hist = FALSE, ...)
}
\arguments{
\item{df}{The data frame as generated by \code{exhaust_design()}.}

\item{est}{A character value to indicate the variable in \code{df} that contains the pont estimate
that you want to plot.}

\item{dchoice}{A character value to indicate the discrete choice that you want to partition
your estimate ridgelines on.}

\item{hist}{Set to true if you want histogram ridges instead of densities. This is
more informative about the exact range of coefficient but does not look as nice.
Sets \code{stat = "binline"} in the \code{\link[ggridges]{geom_density_ridges}} call.
Experiment with parameter \code{scale < 1} to avoid overlapping histograms.}

\item{...}{Additional parameters that are being passed on to  \code{\link[ggridges]{geom_density_ridges}}}
}
\value{
A \code{ggplot} object containing the plot.
}
\description{
Visualizes the variance of an estimate conditional on a selected discrete choice.
}
\details{
See the vignette of the package for further details on how to implement the RDF workflow.
}
\examples{
\dontrun{
  print("Sorry. No examples yet.")
}
}
