% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rdf_spec_curve.R
\name{plot_rdf_spec_curve}
\alias{plot_rdf_spec_curve}
\title{Plot the Specification Curve of your Reseracher Degrees of Freedom}
\usage{
plot_rdf_spec_curve(ests, est, lb = "", ub = "", sample_frac = 0.1,
  sig_color = "red", ribbon_color = "lightblue")
}
\arguments{
\item{ests}{The data frame provided by \code{exhaust_design()}.}

\item{est}{A character string indicating the estimate that you want to plot.}

\item{lb}{A character string indicating the lower bound of the estimate.}

\item{ub}{A character string indicating the upper bound of the estimate.
If both \code{lb} und \code{ub} are present the confindence interval around
the estimate is plotted and the significane of each estimate is assessed.}

\item{sample_frac}{The percentage of the protocols that should be plottted.
If your design produces many degrees of freedom, sampling speeds up plotting
time and makes the plot less clutered.}

\item{sig_color}{Color of significant estimates (only used when \code{lb} and
\code{ub} are present).}

\item{ribbon_color}{Color of the confidence interval (only used when \code{lb} and
\code{ub} are present).}
}
\value{
A plot.
}
\description{
Plots a visulization of the specification curve as introduced by
Simonsohn, Simmons and Nelson.
}
\details{
See the vignette of the package for further details.
}
\examples{
\dontrun{
  print("Sorry. No examples yet.")
}
}
