% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_design_documentation.R
\name{prepare_design_documentation}
\alias{prepare_design_documentation}
\title{Document your research design in a PDF file}
\usage{
prepare_design_documentation(d, output_file, title = "Research Design",
  code = TRUE, rmd_header = paste0("--- \\n", sprintf("title: \\"\%s\\"\\n",
  title), sprintf("date: \\"\%s\\"\\n", Sys.Date()), "output: pdf_document\\n",
  "---\\n\\n"))
}
\arguments{
\item{d}{The design vector containing the research design steps.}

\item{output_file}{The file name to store the PDF file to.}

\item{title}{A title string.}

\item{code}{A logical value indicating whether the R code of the design
steps should be included in the documentation (defaults to TRUE).}

\item{rmd_header}{The character value containing the header of the rmarkdown
file that the function generates. Defaults to a PDF output header}
}
\description{
Uses the documentation block of each step to generate a rmarkdown based
documentation of the research design. Output can be saved to a PDF file
or directly viewed.
}
\examples{
  demo_step <- function(input = NULL, choice = NULL) {
    step_description <- doc(
      "## demo_step",
      "### Content",
      "",
      "This is a demo step. It simply returns two times the value of the choice variable"
    )
    choice_description <- doc(
      "### Choice",
      "",
      "`base_value`: A numerical value that needs to be within the range [0, 10]"
    )
    choice_type <- list(
      list(name = "base_value",
      type = "double",
      valid_min = 0,
      valid_max = 10)
    )
    if (is.null(choice)) return(list(
      step_description = step_description,
      choice_description = choice_description,
      choice_type = choice_type
    )) else check_choice(choice, choice_type)

    return(list(
      data = choice[[1]] * 2,
      protocol = choice
    ))
  }

  design <- "demo_step"
  prepare_design_documentation(design, "my_design.pdf")
}
