% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_design_flow_chart.R
\name{prepare_design_flow_chart}
\alias{prepare_design_flow_chart}
\title{Document your research design in a PDF file}
\usage{
prepare_design_flow_chart(d, file_name = NULL, landscape = FALSE,
  color_step = "red", color_dchoice = "green", color_cchoice = "blue",
  color_result = "yellow")
}
\arguments{
\item{d}{The design vector containing the research design steps.}

\item{file_name}{A file name for an output PDF file relative to the working directory.}

\item{landscape}{Do you want the flow chart to be generated horizontally instead of vertically.
Defaults to FALSE.}

\item{color_step}{The fill color to indicate a design step.}

\item{color_dchoice}{The fill color to indicate a discrete choice.}

\item{color_cchoice}{The fill color to indicate a continuous choice.}

\item{color_result}{The fill color to indicate a result.}
}
\value{
the rendered diagram or, if \code{file_name} is specified, the return code from
  \code{DiagrammeRsvg::export_graph()}
}
\description{
Uses the documentation block of each step to generate a rmarkdown based
documentation of the research design. Output can be saved to a PDF file
or directly viewed.
}
\examples{
  demo_step <- function(input = NULL, choice = NULL) {
    step_description <- doc(
      "## demo_step",
      "### Content",
      "",
      "This is a demo step. It simply returns two times the value of the choice variable"
    )
    choice_description <- doc(
      "### Choice",
      "",
      "`base_value`: A numerical value that needs to be within the range [0, 10]"
    )
    choice_type <- list(
      list(name = "base_value",
      type = "double",
      valid_min = 0,
      valid_max = 10)
    )
    if (is.null(choice)) return(list(
      step_description = step_description,
      choice_description = choice_description,
      choice_type = choice_type
    )) else check_choice(choice, choice_type)

    return(list(
      data = choice[[1]] * 2,
      protocol = choice
    ))
  }

  design <- "demo_step"
  prepare_design_flow_chart(design)
}
