% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_design_power.R
\name{simulate_design_power}
\alias{simulate_design_power}
\title{Use Simulated Data to Test the Power of Your Research Design}
\usage{
simulate_design_power(
  d,
  protocol,
  input_sim_func,
  range_n,
  effect_size,
  runs = 100,
  input_sim_params = NULL
)
}
\arguments{
\item{d}{A character vector of the research design steps function names}

\item{protocol}{A list of \code{choice}s for the \code{step}s of the \code{design}}

\item{input_sim_func}{A function that, when called, will return a simulated input to the first step
for the function. See Details section for further information.}

\item{range_n}{A vector containing the differnt sample sizes for which you wan to estimate the
power of your design.}

\item{effect_size}{The effect size that you want your simulated data to reflect.}

\item{runs}{How many runs do you want to simulate to estimate you power?}

\item{input_sim_params}{If your \code{input_sim_func} needs additional parameters
besides \code{n} and \code{effect_size} you can provide a list of them here.}
}
\value{
A data frame containing the \code{output} of the last \code{step} of the \code{design}
  along with the simulation parameters,
}
\description{
Uses your simulation code, a vector of potential sample sizes and your
a priori assumption of the effect size to estimate the power of your
research design.
}
\details{
The function provided via \code{input_sim_func} needs to take a sample size parameter and an
effect size parameter as the first two input parameters.
See the vignette of the package for additional information on how to implement the RDF analysis
workfow.
}
\examples{
\dontrun{
  print("Sorry. No examples yet.")
}
}
