% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_design.R
\name{source_design}
\alias{source_design}
\title{Source the R Files that Contain the Design Steps}
\usage{
source_design(
  d,
  rel_dir = "code",
  one_file = FALSE,
  one_file_name = "design_steps.R"
)
}
\arguments{
\item{d}{A character vector of the research design steps function names}

\item{rel_dir}{The relative path to your code directory, deafults to "code"}

\item{one_file}{If \code{TRUE}, then all step code is expected to be in a file
with the name \code{one_file_name}. If {FALSE} (the default), each step code
is expected to be in a separate file with the step name as file name.}

\item{one_file_name}{The name of the code file conating the code if
all step code is in one file. Defaults to "design_steps.R".}
}
\value{
The result of the \code{source()} command.
}
\description{
This sources all code files constructed by \code{define_design()}
}
\details{
See the vignette of the package for further details.
}
\examples{
\dontrun{
  print("Sorry. No examples yet.")
}
}
