% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_design.R
\name{test_design}
\alias{test_design}
\title{Test Design for Code Consistency}
\usage{
test_design(d, input = NULL, input_test_code = NULL,
  output_test_code = NULL, ...)
}
\arguments{
\item{d}{The design vector containing the research design steps.}

\item{input}{input parameter of the first step. Can be either data (passed to first step as is)
or a function. If it is a function, the function is called to generate the input data.
If you do not provide \code{input} then only \code{step_description}, \code{choice_description}
and \code{choice_type} of each step are tested for consisitency. If \code{input} is provided,
you also need an R code file to test \code{input} for validity.}

\item{input_test_code}{A path to an R file containg test code to verifiy the integrity of \code{input}
data.}

\item{output_test_code}{A character vector of a length equal to the length of \code{d} containing
paths to test code files for each step's output.}

\item{...}{Are passed to \code{testthat::test_file()}}
}
\value{
The return value \code{testthat::test_file()}
}
\description{
Test whether design steps can be processed and produce valid output for each possible choice.
}
\details{
See the vignette of the package for further details.
}
\examples{
  demo_step <- function(input = NULL, choice = NULL) {
    step_description <- doc(
      "## demo_step",
      "### Content",
      "",
      "This is a demo step. It simply returns two times the value of the choice variable"
    )
    choice_description <- doc(
      "### Choice",
      "",
      "`base_value`: A numerical value that needs to be within the range [0, 10]"
    )
    choice_type <- list(
      list(name = "base_value",
      type = "double",
      valid_min = 0,
      valid_max = 10)
    )
    if (is.null(choice)) return(list(
      step_description = step_description,
      choice_description = choice_description,
      choice_type = choice_type
    )) else check_choice(choice, choice_type)

    return(list(
      data = choice[[1]] * 2,
      protocol = choice
    ))
  }

  design <- "demo_step"
 test_design(design)
}
