% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_choice.R
\name{check_choice}
\alias{check_choice}
\title{Check Design Step Choice}
\usage{
check_choice(choice, choice_type)
}
\arguments{
\item{choice}{a list containing choices}

\item{choice_type}{a list of lists characterizing valid choices. Needs to be specified
by the user for each design step.}
}
\description{
Checks whether a certain choice complies with a provided choice_type.
This helper function is called by the step template after the documentation block
when choice is not \code{NULL}.
}
\examples{
  check_choice(list("my_choice"),
    list(list(name = "achoice",
              type = "character",
              valid_values = c("my_choice", "another_choice"))))
}
