% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rdf_ridges_by_dchoice.R
\name{plot_rdf_ridges_by_dchoice}
\alias{plot_rdf_ridges_by_dchoice}
\title{Plot Estimate Ridge Lines by Discrete Choice}
\usage{
plot_rdf_ridges_by_dchoice(df, est, dchoice, ...)
}
\arguments{
\item{df}{The data frame as generated by \code{exhaust_design()}.}

\item{est}{A character value to indicate the variable in \code{df} that contains the pont estimate
that you want to plot.}

\item{dchoice}{A character value to indicate the discrete choice that you want to partition
your estimate ridgelines on.}

\item{...}{Additional parameters that are being passed on to \code{ggridges::geom_density_redges()}}
}
\value{
A \code{ggplot} object containing the plot.
}
\description{
Visualizes the variance of an estimate conditional on a selected discrete choice.
}
\details{
See the vignette of the package for further details on how to implement the RDF workflow.
}
\examples{
\dontrun{
  print("Sorry. No examples yet.")
}
}
