% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rdf_spec_curve.R
\name{plot_rdf_spec_curve}
\alias{plot_rdf_spec_curve}
\title{Plot the Specification Curve of your Researcher Degrees of Freedom}
\usage{
plot_rdf_spec_curve(ests, est, lb = "", ub = "", sample_frac = 1,
  est_label = est, est_color = "black", est_color_signeg = "yellow",
  est_color_sigpos = "blue", ribbon = nrow(ests) > 30,
  ribbon_color = ifelse(nrow(ests) > 30, "lightblue", "black"),
  pt_size = 0.1, lower_to_upper = 3, highlight = NULL,
  pt_size_highlight = 3)
}
\arguments{
\item{ests}{The data frame provided by \code{\link[rdfanalysis:exhaust_design]{exhaust_design()}}.}

\item{est}{A character string indicating the estimate that you want to plot.}

\item{lb}{A character string indicating the lower bound of the estimate.}

\item{ub}{A character string indicating the upper bound of the estimate.
If both \code{lb} und \code{ub} are present the plot includes a confindence
interval around the estimate and colors the estimate points according to
their significance (see below).}

\item{sample_frac}{The percentage of the protocols that should be plottted.
If your design produces many degrees of freedom, sampling speeds up plotting
time and makes the plot less clutered.}

\item{est_label}{Y-Axis label for the specification curve. Defaults to
\code{est}.}

\item{est_color}{Color of estimates points for specification curve}

\item{est_color_signeg}{Color of significantly negative estimates.}

\item{est_color_sigpos}{Color of significantly positive estimates.}

\item{ribbon}{If TRUE a ribbon displaying the confidence interval is being
plotted. If FALSE, lines are printed instead (looks nicer if you have only
few degrees of freedom to plot). \code{ribbon_color} applies
in either case (only used when \code{lb} and
\code{ub} are present).}

\item{ribbon_color}{Color of the confidence interval (only used when \code{lb} and
\code{ub} are present).}

\item{pt_size}{Point plot size for estimates and choice indicators that are
not highlighted (see below).}

\item{lower_to_upper}{The size of the choice part of the plot, relative to
the specification curve itself.}

\item{highlight}{\code{NULL} or, if you want to highlight certain protocols by
their point size, a data frame containing the choices that identify the
protocols.}

\item{pt_size_highlight}{Point plot size for highlighted estimates and
choice indicators.}
}
\value{
A ggplot object.
}
\description{
Visualizes a result estimate across your researcher degrees of
freedom by plotting its specification curve as introduced by
\href{https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2694998}{Simonsohn, Simmons and Nelson}.
}
\details{
Significance of estimates is only dislayed when \code{lb} and
\code{ub} are present. It is assessed by testing whether
\code{lb} and \code{ub} are of equal sign.

See the vignette of the package for further details.
}
\examples{
\dontrun{
  print("Sorry. No examples yet.")
}
}
