% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_design_documentation.R
\name{prepare_design_documentation}
\alias{prepare_design_documentation}
\title{Document your research design in a PDF file}
\usage{
prepare_design_documentation(d, output_file, title = "Research Design",
  code = TRUE, rmd_header = c("--- ", sprintf("title: \\"\%s\\"",
  title), sprintf("date: \\"\%s\\"", Sys.Date()), "output: pdf_document",
  "---", ""))
}
\arguments{
\item{d}{The design vector containing the research design steps.}

\item{output_file}{The file name to store the PDF file to.}

\item{title}{A title string.}

\item{code}{A logical value indicating whether the R code of the design
steps should be included in the documentation (defaults to TRUE).}

\item{rmd_header}{The character value containing the header of the rmarkdown
file that the function generates. Defaults to a PDF output header}
}
\description{
Uses the documentation block of each step to generate a rmarkdown based
documentation of the research design. Output can be saved to a PDF file
or directly viewed.
}
\examples{
\dontrun{
  print("Sorry. No examples yet.")
}
}
