% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_design.R
\name{test_design}
\alias{test_design}
\title{Test Design for Code Consistency}
\usage{
test_design(
  d,
  input = NULL,
  input_test_code = NULL,
  output_test_code = NULL,
  ...
)
}
\arguments{
\item{d}{A character vector of the research design steps function names}

\item{input}{input parameter of the first step. Can be either data (passed to first step as is)
or a function. If it is a function, the function is called to generate the input data.
If you do not provide \code{input} then only \code{step_description}, \code{choice_description}
and \code{choice_type} of each step are tested for consisitency. If \code{input} is provided,
you also need an R code file to test \code{input} for validity.}

\item{input_test_code}{A path to an R file containg test code to verifiy the integrity of \code{input}
data.}

\item{output_test_code}{A character vector of a length equal to the length of \code{d} containing
paths to test code files for each step's output.}

\item{...}{Are passed to \code{testthat::test_file()}}
}
\value{
The return value \code{testthat::test_file()}
}
\description{
Test whether design steps can be processed and produce valid output for each possible choice.
}
\details{
If called without parameters for \code{input}, only the \code{choice_type} list structure is
  tested for consistency. If \code{input} is provided, data steps are tested across all choices to verify
  that they produce structually valid output without errors or warnings.
  If \code{input_test_code} is provided this test code is used to verify \code{input}.
  If \code{output_test_code} is provided each step's output is tested.
  See the vignette of the package for further details.
}
\examples{
\dontrun{
  print("Sorry. No examples yet.")
}
}
